+++ src/webgl/shaderity_shaders/ClassicSingleShader/ClassicSingleShader.vert
@@ -32,11 +32,11 @@ void main()
 
 /* shaderity: @{mainPrerequisites} */
 
-  mat4 worldMatrix = get_worldMatrix(a_instanceInfo.x);
+  mat4 worldMatrix = get_worldMatrix(uint(a_instanceInfo.x));
   mat4 viewMatrix = get_viewMatrix(cameraSID);
   mat4 projectionMatrix = get_projectionMatrix(cameraSID);
-  mat3 normalMatrix = get_normalMatrix(a_instanceInfo.x);
-  bool isBillboard = get_isBillboard(a_instanceInfo.x);
+  mat3 normalMatrix = get_normalMatrix(uint(a_instanceInfo.x));
+  bool isBillboard = get_isBillboard(uint(a_instanceInfo.x));
 
   // Skeletal
   processGeometry(
@@ -61,11 +61,11 @@ void main()
   v_texcoord_0 = a_texcoord_0;
   v_baryCentricCoord = a_baryCentricCoord.xyz;
 
-  bool visibility = get_isVisible(a_instanceInfo.x);
+  bool visibility = get_isVisible(uint(a_instanceInfo.x));
   if (!visibility)
   {
     gl_Position = vec4(0.0);
   }
 
-  v_shadowCoord = get_depthBiasPV(materialSID, 0) * v_position_inWorld;
+  v_shadowCoord = get_depthBiasPV(materialSID, 0u) * v_position_inWorld;
 }
