+++ src/webgl/shaderity_shaders/EntityUIDOutputSingleShader/EntityUIDOutputSingleShader.vert
@@ -19,10 +19,10 @@ void main()
 {
 /* shaderity: @{mainPrerequisites} */
 
-  mat4 worldMatrix = get_worldMatrix(a_instanceInfo.x);
+  mat4 worldMatrix = get_worldMatrix(uint(a_instanceInfo.x));
   mat4 viewMatrix = get_viewMatrix(cameraSID);
   mat4 projectionMatrix = get_projectionMatrix(cameraSID);
-  mat3 normalMatrix = get_normalMatrix(a_instanceInfo.x);
+  mat3 normalMatrix = get_normalMatrix(uint(a_instanceInfo.x));
 
   // Skeletal
   processGeometry(
