+++ src/webgl/shaderity_shaders/MToon1SingleShader/MToon1SingleShader.vert.glsl
@@ -29,9 +29,9 @@ void main(){
 
   /* shaderity: @{mainPrerequisites} */
 
-  mat4 worldMatrix = get_worldMatrix(a_instanceInfo.x);
+  mat4 worldMatrix = get_worldMatrix(uint(a_instanceInfo.x));
   mat4 viewMatrix = get_viewMatrix(cameraSID);
-  mat3 normalMatrix = get_normalMatrix(a_instanceInfo.x);
+  mat3 normalMatrix = get_normalMatrix(uint(a_instanceInfo.x));
   bool isSkinning = false;
   isSkinning = processGeometry(
     worldMatrix,
@@ -52,12 +52,12 @@ void main(){
   v_normal_inView = vec3(viewMatrix * vec4(v_normal_inWorld, 0.0));
 
 #ifdef RN_MTOON_IS_OUTLINE
-  int outlineWidthType = get_outlineWidthMode(materialSID, 0);
+  int outlineWidthType = get_outlineWidthMode(materialSID, 0u);
   if (outlineWidthType == 0) { // 0 ("none")
     gl_Position = projectionMatrix * viewMatrix * v_position_inWorld;
   } else {
     float worldNormalLength = length(normalMatrix * a_normal);
-    float outlineWidthFactor = get_outlineWidthFactor(materialSID, 0);
+    float outlineWidthFactor = get_outlineWidthFactor(materialSID, 0u);
     vec3 outlineOffset = outlineWidthFactor * worldNormalLength * v_normal_inWorld;
 
     float outlineWidthMultiply = texture(u_outlineWidthMultiplyTexture, a_texcoord_0).g;
