+++ src/webgl/shaderity_shaders/VarianceShadowMapDecodeClassicShader/VarianceShadowMapDecodeClassicShader.vert
@@ -20,8 +20,8 @@ out vec4 v_texcoord_light;
 void main(){
 /* shaderity: @{mainPrerequisites} */
 
-  mat3 normalMatrix = get_normalMatrix(a_instanceInfo.x);
-  mat4 worldMatrix = get_worldMatrix(a_instanceInfo.x);
+  mat3 normalMatrix = get_normalMatrix(uint(a_instanceInfo.x));
+  mat4 worldMatrix = get_worldMatrix(uint(a_instanceInfo.x));
   mat4 viewMatrix = get_viewMatrix(cameraSID);
 
   bool isSkinning = false;
@@ -49,7 +49,7 @@ void main(){
   v_texcoord_0 = a_texcoord_0;
 
   // Shadow mapping
-  mat4 lightViewProjectionMatrix = get_lightViewProjectionMatrix(materialSID, 0);
+  mat4 lightViewProjectionMatrix = get_lightViewProjectionMatrix(materialSID, 0u);
   v_projPosition_from_light = lightViewProjectionMatrix * v_position_inWorld;
 
   // Following tMatrix is based on https://wgld.org/d/webgl/w051.html
