+++ src/webgl/shaderity_shaders/nodes/ProjectionMatrix.vert
@@ -1,5 +1,5 @@
 
 void projectionMatrix(out mat4 outValue) {
-  float cameraSID = u_currentComponentSIDs[/* shaderity: @{WellKnownComponentTIDs.CameraComponentTID}*/];
+  uint cameraSID = uint(u_currentComponentSIDs[/* shaderity: @{WellKnownComponentTIDs.CameraComponentTID}*/]);
   outValue = get_projectionMatrix(cameraSID);
 }
